crypto/flippin-auth/challenge/src/requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycryptodome==3.17
Werkzeug==2.2.2
